+++ lithops/serverless/backends/azure_functions/azure_functions.py
@@ -304,14 +304,14 @@ class AzureFunctionAppBackend:
 
             return resp_text
 
-    def get_runtime_key(self, runtime_name, runtime_memory):
+    def get_runtime_key(self, runtime_name, runtime_memory, version=__version__):
         """
         Method that creates and returns the runtime key.
         Runtime keys are used to uniquely identify runtimes within the storage,
         in order to know which runtimes are installed and which not.
         """
-        function_name = self._format_function_name(runtime_name, runtime_memory)
-        runtime_key = os.path.join(self.name, __version__, function_name)
+        function_name = self._format_function_name(runtime_name, runtime_memory, version)
+        runtime_key = os.path.join(self.name, version, function_name)
 
         return runtime_key