+++ lithops/serverless/backends/gcp_functions/gcp_functions.py
@@ -382,9 +382,9 @@ class GCPFunctionsBackend:
         else:
             raise Exception(f'Error at retrieving runtime meta: {response}')
 
-    def get_runtime_key(self, runtime_name, runtime_memory):
-        function_name = self._format_function_name(runtime_name, runtime_memory)
-        runtime_key = os.path.join(self.name, __version__, self.project_name, self.region, function_name)
+    def get_runtime_key(self, runtime_name, runtime_memory, version=__version__):
+        function_name = self._format_function_name(runtime_name, runtime_memory, version)
+        runtime_key = os.path.join(self.name, version, self.project_name, self.region, function_name)
         logger.debug(f'Runtime key: {runtime_key}')
 
         return runtime_key