+++ lithops/serverless/backends/openwhisk/openwhisk.py
@@ -51,7 +51,7 @@ class OpenWhiskBackend:
         logger.debug(f"Set OpenWhisk Insecure to {self.insecure}")
 
         self.user_key = self.api_key[:5]
-        self.package = f'lithops_v{__version__}_{self.user_key}'
+        self.package = f'lithops_{self.user_key}'
 
         self.cf_client = OpenWhiskClient(
             endpoint=self.endpoint,
@@ -159,7 +159,7 @@ class OpenWhiskBackend:
 
         packages = self.cf_client.list_packages()
         for pkg in packages:
-            if pkg['name'].startswith('lithops') and pkg['name'].endswith(self.user_key):
+            if pkg['name'] == self.package:
                 actions = self.cf_client.list_actions(pkg['name'])
                 for action in actions:
                     version, image_name, memory = self._unformat_function_name(action['name'])
@@ -178,14 +178,14 @@ class OpenWhiskBackend:
 
         return activation_id
 
-    def get_runtime_key(self, docker_image_name, runtime_memory):
+    def get_runtime_key(self, docker_image_name, runtime_memory, version=__version__):
         """
         Method that creates and returns the runtime key.
         Runtime keys are used to uniquely identify runtimes within the storage,
         in order to know which runtimes are installed and which not.
         """
-        action_name = self._format_function_name(docker_image_name, runtime_memory)
-        runtime_key = os.path.join(self.name, __version__, self.namespace, action_name)
+        action_name = self._format_function_name(docker_image_name, runtime_memory, version)
+        runtime_key = os.path.join(self.name, version, self.namespace, action_name)
 
         return runtime_key